add a migration that we missed
I was developing locally and saw the "Your models have changes that
are not yet reflected in a migration, and so won't be applied."
message. Investigating, it looks like django wants a migration to
apply the `on_delete=DO_NOTHING` that we had to add for Django 2.x
compatability in 2fbd23fe5de2abe61a608c22e3f59e337eef6c4a.